--- frontend/src/App.css
@@ -1,11 +1,0 @@
-:root{
-  --blue: #377dff;
-  --dark-blue: #21325b;
-  --violet: #b253f6;
-  --grey-back: #e7e7ec;
-  --dark-text: #2f2d51;
-  --light-text: #77838f;
-  --light-orange: #FFC78B;
-  --light-violet: #B8B2FD;
-  --dark-gray-70:rgba(47, 45, 81, 0.7);
-}
