+++ frontend/src/App.js
@@ -1,13 +1,10 @@
-import './App.css';
-import FeaturedCourses from './components/contentBlocks/11home3block/FeaturedCourses';
-import CourseCard from './components/contentBlocks/CommonItems/CourseCard/CourseCard';
+import '../src/components/contentBlocks/commonCss/styles.css';
+import FeaturedCourses from './components/contentBlocks/11home3block';
 
 function App() {
   return (
       <div>
-
         <FeaturedCourses/>
-
       </div>       
   );
 }
