+++ frontend/src/components/contentBlocks/CommonItems/CourseCard/index.js
@@ -1,4 +1,4 @@
-import './CourseCard.css';
+import './styles.css';
 
 const CourseCard = ({
     cardId,
