+++ frontend/src/components/contentBlocks/CommonItems/CourseCard/styles.css
